Bug 29496: (bug 27526 follow-up) Fix item form validation
The validation of the forms were blocked with "X item mandatory fields
empty" when at least one dropdown list subfield was marked as mandatory.

We need to add the 'input_marceditor' class to the select (does it
actually make sense? select vs input...)

Caused by
  commit 6ed29bccef590a1bc5e4cea175b41b8024483320
  Bug 27526: Fix mandatory and important checks
Which lamentably failed as it was stating:
"Using .input_marceditor let us fix the additem.tt form and prevent to break the other ones"

Signed-off-by: Marion Durand <marion.durand@biblibre.com>

Signed-off-by: Marcel de Rooy <m.de.rooy@rijksmuseum.nl>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>