Bug 29516: Remove dependency on IO::Scalar
IO::Scalar is used in Koha::App::Plugin::CGIBinKoha to create a
filehandle tied to a scalar.
Perl has this feature built in since 5.8.0 so IO::Scalar is not needed

I'm not sure how to test this, since it should behave exactly the same
with or without the patch, but we can at least verify that it doesn't
introduce encoding issues.

Test plan:
1. Run `morbo bin/intranet`
2. Go to http://localhost:3000
3. Create a bibliographic record with some non-latin characters (try
   emoji characters for instance). Verify that there is no visible
   encoding issues.
4. Save this record as MARCXML and re-import it as a new record using
   Tools » Stage MARC records for import. Verify that the new record has
   no visible encoding issues
5. `git grep IO::Scalar` should not return any result

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>
Signed-off-by: Fridolin Somers <fridolin.somers@biblibre.com>