Bug 29528: Breadcrumbs on HTML customizations take you to news
This patch removes the "Additional contents" breadcrumb when working
with news items or HTML customizations. Since news and HTML
customizations are separate links on the Tools home page there's no
reason to have the breadcrumbs imply the two sections are connected in
any way. We already have the "See News" link, for example, for switching
quickly between the two areas.

To test, apply the patch and go to Tools -> News.

- The breadcrumbs should be: Home > Tools > News
  - Click "New entry." The breadcrumbs should be: Home > Tools > News >
    Add additional content (News). Clicking "News" should take you to
    the correct page.
- Go to Tools -> HTML customizations.
  - The breadcrumbs should be Home >  Tools > HTML customizations.
  - Click "New entry." The breadcrumbs should be: Home > Tools >  HTML
    customizations > Add additional content (HTML customizations).
    Clicking "HTML customizations" should take you to the correct page.

Signed-off-by: Barbara Johnson <barbara.johnson@bedfordtx.gov>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>
Signed-off-by: Fridolin Somers <fridolin.somers@biblibre.com>